requirements.txt
pyyaml>=6.0
requests>=2.31.0
openai>=1.0.0
html2text>=2020.1.16
trafilatura>=2.0.0
chardet>=5.0.0

# Web Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
aiohttp>=3.9.0

# GUI Dependencies
PySide6>=6.5.0
psutil>=5.9.0
pytest-qt>=4.2.0
